cf_tmp_example/requirements.txt
google-cloud-storage
pygithub
pyyaml